move DomainCoreApi under sp_domains to avoid cyclic deps
