+++ crates/sp-domains-fraud-proof/src/tests.rs
@@ -2,7 +2,7 @@ use crate::test_ethereum_tx::{
     EIP1559UnsignedTransaction, EIP2930UnsignedTransaction, LegacyUnsignedTransaction,
 };
 use codec::Encode;
-use domain_runtime_primitives::{Balance, CheckExtrinsicsValidityError, DomainCoreApi};
+use domain_runtime_primitives::{Balance, CheckExtrinsicsValidityError};
 use domain_test_service::evm_domain_test_runtime::{
     Runtime as TestRuntime, RuntimeCall, UncheckedExtrinsic as RuntimeUncheckedExtrinsic,
 };
@@ -23,6 +23,7 @@ use sp_api::{ApiExt, ProvideRuntimeApi, TransactionOutcome};
 use sp_core::crypto::AccountId32;
 use sp_core::ecdsa::Pair;
 use sp_core::{keccak_256, Pair as _, H160, H256, U256};
+use sp_domains::core_api::DomainCoreApi;
 use sp_runtime::traits::{Extrinsic, Zero};
 use sp_runtime::transaction_validity::{InvalidTransaction, TransactionValidityError};
 use sp_runtime::OpaqueExtrinsic;
@@ -379,7 +380,7 @@ async fn storage_change_of_the_same_runtime_instance_should_perserved_cross_runt
                 best_hash,
                 vec![
                     transfer_to_charlie_with_big_tip_1.clone().into(),
-                    transfer_to_charlie_with_big_tip_2.clone().into()
+                    transfer_to_charlie_with_big_tip_2.clone().into(),
                 ],
                 best_number,
                 best_hash,
@@ -389,7 +390,7 @@ async fn storage_change_of_the_same_runtime_instance_should_perserved_cross_runt
             extrinsic_index: 1,
             transaction_validity_error: TransactionValidityError::Invalid(
                 InvalidTransaction::Payment
-            )
+            ),
         })
     );
 
@@ -411,7 +412,7 @@ async fn storage_change_of_the_same_runtime_instance_should_perserved_cross_runt
             extrinsic_index: 0,
             transaction_validity_error: TransactionValidityError::Invalid(
                 InvalidTransaction::Future
-            )
+            ),
         })
     );
 
@@ -460,7 +461,7 @@ async fn storage_change_of_the_same_runtime_instance_should_perserved_cross_runt
                 best_hash,
                 vec![transfer_with_big_tip_1.clone().into()],
                 best_number,
-                best_hash
+                best_hash,
             )
             .unwrap()
             .is_ok());
@@ -491,7 +492,7 @@ async fn storage_change_of_the_same_runtime_instance_should_perserved_cross_runt
                     best_hash,
                     vec![transfer_with_big_tip_3.clone().into()],
                     best_number,
-                    best_hash
+                    best_hash,
                 )
                 .unwrap(),
             if commit_mode {
