+++ crates/subspace-malicious-operator/src/malicious_bundle_producer.rs
@@ -3,7 +3,6 @@ use domain_client_operator::domain_bundle_producer::DomainBundleProducer;
 use domain_client_operator::domain_bundle_proposer::DomainBundleProposer;
 use domain_client_operator::{OpaqueBundleFor, OperatorSlotInfo};
 use domain_runtime_primitives::opaque::Block as DomainBlock;
-use domain_runtime_primitives::DomainCoreApi;
 use frame_system_rpc_runtime_api::AccountNonceApi;
 use futures::{Stream, StreamExt, TryFutureExt};
 use pallet_domains::OperatorConfig;
@@ -20,6 +19,7 @@ use sp_consensus_slots::Slot;
 use sp_consensus_subspace::FarmerPublicKey;
 use sp_core::crypto::UncheckedFrom;
 use sp_core::Get;
+use sp_domains::core_api::DomainCoreApi;
 use sp_domains::{BundleProducerElectionApi, DomainId, DomainsApi, OperatorId, OperatorPublicKey};
 use sp_keyring::Sr25519Keyring;
 use sp_keystore::KeystorePtr;
