+++ domains/client/block-preprocessor/src/lib.rs
@@ -18,11 +18,11 @@ use crate::inherents::is_runtime_upgraded;
 use crate::xdm_verifier::is_valid_xdm;
 use codec::Encode;
 use domain_runtime_primitives::opaque::AccountId;
-use domain_runtime_primitives::DomainCoreApi;
 use sc_client_api::BlockBackend;
 use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use sp_core::H256;
+use sp_domains::core_api::DomainCoreApi;
 use sp_domains::extrinsics::deduplicate_and_shuffle_extrinsics;
 use sp_domains::{
     DomainId, DomainsApi, ExecutionReceipt, ExtrinsicDigest, HeaderHashingFor, InboxedBundle,
