+++ domains/client/domain-operator/src/fraud_proof.rs
@@ -3,13 +3,14 @@ use crate::ExecutionReceiptFor;
 use codec::{Decode, Encode};
 use domain_block_builder::{BlockBuilder, RecordProof};
 use domain_runtime_primitives::opaque::AccountId;
-use domain_runtime_primitives::{CheckExtrinsicsValidityError, DomainCoreApi};
+use domain_runtime_primitives::CheckExtrinsicsValidityError;
 use sc_client_api::{AuxStore, BlockBackend, ProofProvider};
 use sp_api::{ApiExt, ProvideRuntimeApi};
 use sp_blockchain::HeaderBackend;
 use sp_core::traits::CodeExecutor;
 use sp_core::H256;
 use sp_domain_digests::AsPredigest;
+use sp_domains::core_api::DomainCoreApi;
 use sp_domains::proof_provider_and_verifier::StorageProofProvider;
 use sp_domains::{DomainId, DomainsApi, ExtrinsicDigest, HeaderHashingFor, InvalidBundleType};
 use sp_domains_fraud_proof::execution_prover::ExecutionProver;
