+++ domains/primitives/runtime/src/lib.rs
@@ -27,16 +27,13 @@ use frame_system::limits::{BlockLength, BlockWeights};
 use parity_scale_codec::{Decode, Encode};
 use scale_info::TypeInfo;
 use serde::{Deserialize, Serialize};
-use sp_runtime::generic::{Era, UncheckedExtrinsic};
-use sp_runtime::traits::{
-    Block as BlockT, CheckedAdd, Convert, IdentifyAccount, NumberFor, Verify,
-};
+use sp_runtime::generic::UncheckedExtrinsic;
+use sp_runtime::traits::{Convert, IdentifyAccount, Verify};
 use sp_runtime::transaction_validity::TransactionValidityError;
-use sp_runtime::{Digest, MultiAddress, MultiSignature, Perbill};
+use sp_runtime::{MultiAddress, MultiSignature, Perbill};
 use sp_std::vec::Vec;
 use sp_weights::Weight;
-use subspace_core_primitives::U256;
-use subspace_runtime_primitives::{Moment, SHANNON};
+use subspace_runtime_primitives::SHANNON;
 
 /// Alias to 512-bit hash when used in the context of a transaction signature on the chain.
 pub type Signature = MultiSignature;
@@ -202,42 +199,6 @@ pub struct DecodeExtrinsicError(pub String);
 pub const CHECK_EXTRINSICS_AND_DO_PRE_DISPATCH_METHOD_NAME: &str =
     "DomainCoreApi_check_extrinsics_and_do_pre_dispatch";
 
-#[derive(Clone, Debug, Decode, Default, Encode, Eq, PartialEq, TypeInfo)]
-pub struct BlockFees<Balance> {
-    /// The consensus chain storage fee
-    pub consensus_storage_fee: Balance,
-    /// The domain execution fee including the storage and compute fee on domain chain,
-    /// tip, and the XDM reward.
-    pub domain_execution_fee: Balance,
-}
-
-/// Type that holds the transfers(in/out) for a given chain.
-#[derive(Debug, Decode, Encode, TypeInfo, PartialEq, Eq, Clone, Default)]
-pub struct Transfers<Balance> {
-    /// Total transfers that came into the domain.
-    pub transfers_in: Balance,
-    /// Total transfers that went out of the domain.
-    pub transfers_out: Balance,
-}
-
-impl<Balance> BlockFees<Balance>
-where
-    Balance: CheckedAdd,
-{
-    pub fn new(domain_execution_fee: Balance, consensus_storage_fee: Balance) -> Self {
-        BlockFees {
-            consensus_storage_fee,
-            domain_execution_fee,
-        }
-    }
-
-    /// Returns the total fees that was collected and burned on the Domain.
-    pub fn total_fees(&self) -> Option<Balance> {
-        self.consensus_storage_fee
-            .checked_add(&self.domain_execution_fee)
-    }
-}
-
 /// Opaque types. These are used by the CLI to instantiate machinery that don't need to know
 /// the specifics of the runtime. They can then be made to be agnostic over specific formats
 /// of data like extrinsics, allowing for them to continue syncing the network through upgrades
@@ -257,77 +218,6 @@ pub mod opaque {
     pub type BlockId = generic::BlockId<Block>;
     /// Opaque account identifier type.
     pub type AccountId = Vec<u8>;
-}
-
-sp_api::decl_runtime_apis! {
-    /// Base API that every domain runtime must implement.
-    pub trait DomainCoreApi {
-        /// Extracts the optional signer per extrinsic.
-        fn extract_signer(
-            extrinsics: Vec<<Block as BlockT>::Extrinsic>,
-        ) -> Vec<(Option<opaque::AccountId>, <Block as BlockT>::Extrinsic)>;
-
-        fn is_within_tx_range(
-            extrinsic: &<Block as BlockT>::Extrinsic,
-            bundle_vrf_hash: &U256,
-            tx_range: &U256,
-        ) -> bool;
-
-        /// Returns the intermediate storage roots in an encoded form.
-        fn intermediate_roots() -> Vec<[u8; 32]>;
-
-        /// Returns the storage root after initializing the block.
-        fn initialize_block_with_post_state_root(header: &<Block as BlockT>::Header) -> Vec<u8>;
-
-        /// Returns the storage root after applying the extrinsic.
-        fn apply_extrinsic_with_post_state_root(extrinsic: <Block as BlockT>::Extrinsic) -> Vec<u8>;
-
-        /// Returns an encoded extrinsic aiming to upgrade the runtime using given code.
-        fn construct_set_code_extrinsic(code: Vec<u8>) -> Vec<u8>;
-
-        /// Returns an encoded extrinsic to set timestamp.
-        fn construct_timestamp_extrinsic(moment: Moment) -> Block::Extrinsic;
-
-        /// Returns an encoded extrinsic to set domain transaction byte fee.
-        fn construct_consensus_chain_byte_fee_extrinsic(consensus_chain_byte_fee: Balance) -> Block::Extrinsic;
-
-        /// Returns true if the extrinsic is an inherent extrinsic.
-        fn is_inherent_extrinsic(extrinsic: &<Block as BlockT>::Extrinsic) -> bool;
-
-        /// Checks the validity of array of extrinsics + pre_dispatch
-        /// returning failure on first extrinsic that fails runtime call.
-        /// IMPORTANT: Change `CHECK_EXTRINSICS_AND_DO_PRE_DISPATCH_METHOD_NAME` constant when this method name is changed
-        fn check_extrinsics_and_do_pre_dispatch(uxts: Vec<<Block as BlockT>::Extrinsic>, block_number: NumberFor<Block>,
-            block_hash: <Block as BlockT>::Hash) -> Result<(), CheckExtrinsicsValidityError>;
-
-        /// Decodes the domain specific extrinsic from the opaque extrinsic.
-        fn decode_extrinsic(
-            opaque_extrinsic: sp_runtime::OpaqueExtrinsic,
-        ) -> Result<<Block as BlockT>::Extrinsic, DecodeExtrinsicError>;
-
-        /// Returns extrinsic Era if present.
-        fn extrinsic_era(
-          extrinsic: &<Block as BlockT>::Extrinsic
-        ) -> Option<Era>;
-
-        /// Returns the extrinsic weight.
-        fn extrinsic_weight(ext: &Block::Extrinsic) -> Weight;
-
-        /// The accumulated transaction fee of all transactions included in the block.
-        fn block_fees() -> BlockFees<Balance>;
-
-        /// Returns the block digest.
-        fn block_digest() -> Digest;
-
-        /// Returns the consumed weight of the block.
-        fn block_weight() -> Weight;
-
-        /// Returns the transfers for this domain in the block.
-        fn transfers() -> Transfers<Balance>;
-
-        /// Returns the storage key for the Transfers on Domain.
-        fn transfers_storage_key() -> Vec<u8>;
-    }
 }
 
 #[cfg(test)]
