+++ domains/runtime/evm/src/lib.rs
@@ -17,7 +17,7 @@ pub use domain_runtime_primitives::{
     EXISTENTIAL_DEPOSIT, MAXIMUM_BLOCK_WEIGHT,
 };
 use domain_runtime_primitives::{
-    CheckExtrinsicsValidityError, DecodeExtrinsicError, Transfers, SLOT_DURATION,
+    CheckExtrinsicsValidityError, DecodeExtrinsicError, SLOT_DURATION,
 };
 use fp_account::EthereumSignature;
 use fp_self_contained::{CheckedSignature, SelfContainedCall};
@@ -41,7 +41,7 @@ use pallet_transporter::EndpointHandler;
 use sp_api::impl_runtime_apis;
 use sp_core::crypto::KeyTypeId;
 use sp_core::{Get, OpaqueMetadata, H160, H256, U256};
-use sp_domains::DomainId;
+use sp_domains::{DomainId, Transfers};
 use sp_messenger::endpoint::{Endpoint, EndpointHandler as EndpointHandlerT, EndpointId};
 use sp_messenger::messages::{
     BlockInfo, BlockMessagesWithStorageKey, ChainId, CrossDomainMessage,
@@ -856,7 +856,7 @@ impl_runtime_apis! {
         }
     }
 
-    impl domain_runtime_primitives::DomainCoreApi<Block> for Runtime {
+    impl sp_domains::core_api::DomainCoreApi<Block> for Runtime {
         fn extract_signer(
             extrinsics: Vec<<Block as BlockT>::Extrinsic>,
         ) -> Vec<(Option<opaque::AccountId>, <Block as BlockT>::Extrinsic)> {
@@ -959,7 +959,7 @@ impl_runtime_apis! {
             ext.get_dispatch_info().weight
         }
 
-        fn block_fees() -> domain_runtime_primitives::BlockFees<Balance> {
+        fn block_fees() -> sp_domains::BlockFees<Balance> {
             BlockFees::collected_block_fees()
         }
